backend/requirements.txt
fastapi
uvicorn[standard]
crawl4ai
pydantic
pydantic-settings
python-dotenv
groq
litellm
exa_py
joblib
sentence-transformers
numpy
nltk